add app version, rm heroic version
