+++ src/frontend/screens/Settings/sections/GeneralSettings/index.tsx
@@ -18,12 +18,15 @@ import {
   UseDarkTrayIcon,
   WinePrefixesBasePath
 } from '../../components'
+import AppVersion from 'frontend/components/UI/AppVersion'
 
 export default function GeneralSettings() {
   const { t } = useTranslation()
 
   return (
     <>
+      <AppVersion />
+
       <div className="settingSubheader settingsSectionHeader title">
         {t('settings.navbar.general')}
       </div>
